\name{00.dse.Intro}
\alias{00.dse.Intro}
\docType{package}

\title{Dynamic Systems Estimation - Multivariate Time Series Package}
\description{Functions for multivariate time series modeling} 
\details{ See \code{\link{dse-package}} ( in the help system use
   package?dse or ?"dse-package") for an overview.
}
\keyword{package}



