\name{is.forecastCovEstimatorsWRTdata.subsets}
\alias{is.forecastCovEstimatorsWRTdata.subsets}
\title{Check Inheritance}
\description{Check inheritance.}
\usage{
    is.forecastCovEstimatorsWRTdata.subsets(obj)
    }
\arguments{
    \item{obj}{Any object.}
    }
\value{logical}
\details{
    This tests if an object isnherits 
    from \code{forecastCovEstimatorsWRTdata.subsets}. This type of object
    code be generated in different ways but the only current example
    is \code{stripMine}.
    }
\seealso{
    \code{\link{stripMine}}
    }
\concept{DSE}
\keyword{ts}

