\name{nseries.featherForecasts}
\alias{nseries.featherForecasts}

\title{Number of Series}
\description{Return the number of series.}
\usage{
    \method{nseries}{featherForecasts}(x)
    }

\arguments{
    \item{x}{A featherForecasts object.}
    }
\value{An integer.}
\details{
    See the generic method.
    }
\keyword{programming}
\keyword{utilities}
\keyword{ts}


