% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dse.R
\name{parse_tablename}
\alias{parse_tablename}
\title{Parse tablename}
\usage{
parse_tablename(tablename)
}
\arguments{
\item{tablename}{The table name in schema.table or catalog.schema.table format. If catalog name is undefined, it will default to "devcatalog"}
}
\value{
Correctly formatted table name in catalog.schema.table format for DSE
}
\description{
Format the tablename in the catalog.schema.table format
}
\examples{
table <- parse_tablename("myschema.mytable")
table2 <- parse_tablename("mycatalog.myschema.mytable")
}
