\name{print.TSdata}
\alias{print.TSdata}

\title{Print Specific Methods}
\description{See the generic function description.}
\usage{
    \method{print}{TSdata}(x, ...)
    }
\arguments{
    \item{x}{An object of class TSdata.}
    \item{...}{arguments to be passed to other methods.}
}
\seealso{
    \code{\link{print}}
    \code{\link{summary}}
}

\concept{DSE}
\keyword{ts}

