% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dse.R
\name{read_data}
\alias{read_data}
\title{Read data from DSE table}
\usage{
read_data(tablename)
}
\arguments{
\item{tablename}{The table name in schema.table or catalog.schema.table format}
}
\value{
Spark dataframe of requested table
}
\description{
Read data from DSE table
}
\examples{
table <- read_data("myschema.mytable")
}
