% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dse.R
\name{save_data}
\alias{save_data}
\title{Save data to DSE table}
\usage{
save_data(df, tablename, mode = "error")
}
\arguments{
\item{df}{The dataframe to save, of SparkR dataframe or R dataframe format}

\item{tablename}{The table name in schema.table or catalog.schema.table format}
}
\description{
Save data to DSE table
}
\examples{
table <- read_data("myschema.mytable")
}
