\name{seriesNamesInput.forecast}
\alias{seriesNamesInput.forecast}
\alias{seriesNamesInput.featherForecasts}

\alias{seriesNamesOutput.forecast}
\alias{seriesNamesOutput.featherForecasts}

\title{TS Input and Output Specific Methods}
\description{See the generic function description.}
\usage{
    \method{seriesNamesInput}{forecast}(x)
    \method{seriesNamesInput}{featherForecasts}(x)

    \method{seriesNamesOutput}{forecast}(x)
    \method{seriesNamesOutput}{featherForecasts}(x)
}
\arguments{
    \item{x}{an object from which to extract the names of the input or output
      series.}
}

\concept{DSE}
\keyword{ts}


