\name{testEqual.ARMA}
\alias{testEqual.ARMA}
\alias{testEqual.SS}
\alias{testEqual.TSdata}
\alias{testEqual.TSmodel}
\alias{testEqual.TSestModel}

\title{Specific Methods for Testing Equality}
\description{See the generic function description.}
\usage{
    \method{testEqual}{ARMA}(obj1, obj2, fuzz=0)
    \method{testEqual}{SS}(obj1, obj2, fuzz=0)
    \method{testEqual}{TSdata}(obj1, obj2, fuzz=1e-16)
    \method{testEqual}{TSmodel}(obj1, obj2, fuzz=0)
    \method{testEqual}{TSestModel}(obj1, obj2, fuzz=0)
    }
\arguments{
    \item{obj1}{see generic method.}
    \item{obj2}{see generic method.}
    \item{fuzz}{see generic method.}
}

\seealso{\code{\link[tframe]{testEqual}} }

\concept{DSE}
\keyword{ts}

